Added production and development configs. Switched from gunicorn to socketio.run with nginx reverse proxy
