+++ docker-compose.development.yml
@@ -1,0 +1,14 @@
+version: '3.7'
+
+services:
+  flask:
+    volumes:
+      - ./server:/app
+    ports:
+      - 5000:5000
+  frontend:
+    volumes:
+      - ./frontend:/app
+      - /app/node_modules
+    ports:
+      - 8080:8080
