+++ docker-compose.production.yml
@@ -1,0 +1,14 @@
+version: '3.7'
+
+services:
+  frontend:
+    build:
+      context: ./frontend
+      dockerfile: ./Dockerfile.production
+    ports:
+      - 80:80
+    links:
+      - flask
+    depends_on:
+      - flask
+    # restart: unless-stopped
