+++ frontend/Dockerfile
@@ -11,3 +11,4 @@ COPY . .
 ENV VUE_APP_FLASK_HOST flask
 ENV HOST 0.0.0.0
 EXPOSE 8080
+CMD ["npm", "run", "serve"]
