+++ frontend/deploy/prod.nginx.conf
@@ -1,0 +1,42 @@
+#/etc/nginx/sites-available/codenames
+server {
+    # proxy for socket.io
+    location /socket.io {
+        include proxy_params;
+        proxy_http_version 1.1;
+        proxy_redirect off;
+        proxy_buffering off;
+        proxy_set_header Upgrade $http_upgrade;
+        proxy_set_header Origin "";
+        proxy_set_header Connection "Upgrade";
+        proxy_pass http://flask:5000/socket.io;
+    }
+
+    location /stats {
+        include proxy_params;
+        proxy_pass http://flask:5000/stats;
+    }
+
+    location /static/ {
+        root    /usr/share/nginx/html;
+    }
+
+    location / {
+        root    /usr/share/nginx/html;
+        try_files $uri $uri/ /index.html;
+    }
+
+    server_name codewords.tv;
+    listen 80;
+    listen [::]:80;
+}
+
+# server {
+#     # http redirect to https
+#     if ($host = codewords.tv) {
+#         return 301 https://$host$request_uri;
+#     }
+#     listen 80;
+#     server_name codewords.tv;
+#     return 404;
+# }
