+++ frontend/package.json
@@ -24,22 +24,22 @@
     "vue-resize-text": "^0.1.1",
     "vue-router": "^3.1.6",
     "vue-socket.io": "~3.0.7",
-    "vuetify": "^2.2.20",
+    "vuetify": "^2.2.21",
     "vuex": "^3.1.3",
     "vuex-persistedstate": "^3.0.1",
     "vuex-router-sync": "^5.0.0"
   },
   "devDependencies": {
     "@babel/preset-stage-2": "7.8.3",
-    "@vue/cli-plugin-babel": "^4.3.0",
-    "@vue/cli-plugin-eslint": "^4.3.0",
-    "@vue/cli-plugin-pwa": "^4.3.0",
-    "@vue/cli-plugin-unit-jest": "^4.3.0",
-    "@vue/cli-service": "^4.3.0",
+    "@vue/cli-plugin-babel": "^4.3.1",
+    "@vue/cli-plugin-eslint": "^4.3.1",
+    "@vue/cli-plugin-pwa": "^4.3.1",
+    "@vue/cli-plugin-unit-jest": "^4.3.1",
+    "@vue/cli-service": "^4.3.1",
     "@vue/eslint-config-airbnb": "^5.0.2",
-    "@vue/test-utils": "^1.0.0-beta.32",
+    "@vue/test-utils": "^1.0.0-beta.33",
     "babel-core": "7.0.0-bridge.0",
-    "babel-jest": "^25.2.6",
+    "babel-jest": "^25.3.0",
     "node-sass": "^4.13.1",
     "sass-loader": "^8.0.2",
     "vue-template-compiler": "^2.6.11"
