+++ frontend/src/registerServiceWorker.js
@@ -15,7 +15,6 @@ if (process.env.NODE_ENV === 'production') {
     },
     updated () {
       console.log('New content is available; please refresh.')
-      window.location.reload(true)
     },
     offline () {
       console.log('No internet connection found. App is running in offline mode.')
