+++ server/Dockerfile
@@ -8,4 +8,4 @@ RUN pip install -r requirements.txt
 COPY . .
 ENV HOST 0.0.0.0
 EXPOSE 5000
-# CMD ["python", "__init__.py"]
+CMD ["python", "app.py"]
