+++ server/requirements.txt
@@ -6,7 +6,7 @@ dnspython==1.16.0
 dominate==2.5.1
 enum-compat==0.0.3
 enum34==1.1.10
-eventlet==0.25.1
+eventlet==0.25.2
 Flask==1.1.2
 Flask-Bootstrap==3.3.7.1
 Flask-SocketIO==4.2.1
@@ -34,4 +34,4 @@ urllib3==1.25.8
 visitor==0.1.3
 WebOb==1.8.6
 Werkzeug==1.0.1
-wrapt==1.11.2
+wrapt==1.12.1
